requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
docopt==0.6.2
Flask==1.1.1
Flask-Caching==1.8.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.1
Pillow==7.0.0
pymongo==3.10.0
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
tqdm==4.40.2
urllib3==1.25.7
Werkzeug==0.16.0
wordcloud==1.6.0
yarg==0.1.9